refactor: modularize hero preview components and update responsive layout containers for documentation pages.
